+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
-- Fix errors with renames failing across disks by falling back to a copy/delete. [#30](https://github.com/andrewkroh/gvm/pull/30)
+- Fix errors with renames failing across disks by falling back to a copy/delete. [#31](https://github.com/andrewkroh/gvm/pull/31)
 
 ## [0.2.3]